requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-braces==1.11.0
django-filter==1.1.0
django-filters==0.2.1
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
